Fix executable resubmission behavior on instance termination
(computer.getOfflineCause() instanceof OfflineCause.ChannelTermination) does not always hold after an unexpected instance termination.
As reported in issue #121, the offline cause is sometimes simply hudson.slaves.OfflineCause$SimpleOfflineCause; this led to executables only being occasionally resubmitted.

This commit makes it so that active executables are always resubmitted regardless of the offline cause.
The assumption is that active executables that fail due to an instance going offline should always be automatically rescheduled (unless disableTaskResubmit is specified)